Fix PC relog issue.
When NWN loads a character that is relogging in it first creates a whole new creature,
loads its items and runs related events, applies item properties, etc.  Then when the
creature object is constructed it reverts the object id to its original first login value.

This creates serious havoc when attempting to access the cache during the brief
period before the object id is reset, since as far as Solstice is concerned it is a new
creature.

So I've added a look up table indexed by player name and character name to find
the canonical object id.  Sadly this adds some overhead every time a PC object is
looked up.  It should probably be re-visited to improve that.